move: vo 패키지를 domain 으로 이동
