+++ src/main/java/subway/application/LineService.java
@@ -10,7 +10,7 @@ import subway.dao.StationDao;
 import subway.domain.Line;
 import subway.domain.Section;
 import subway.domain.Station;
-import subway.vo.StationPair;
+import subway.domain.vo.StationPair;
 import subway.domain.Stations;
 import subway.ui.dto.LineRequest;
 import subway.ui.dto.LineResponse;
