+++ src/main/java/subway/domain/Stations.java
@@ -10,7 +10,7 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import subway.exception.IllegalStationsException;
-import subway.vo.StationPair;
+import subway.domain.vo.StationPair;
 
 public class Stations {
