+++ src/test/java/subway/domain/StationsTest.java
@@ -8,7 +8,7 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import subway.exception.IllegalStationsException;
-import subway.vo.StationPair;
+import subway.domain.vo.StationPair;
 
 class StationsTest {
